M2C-125
Implement business logic to convert a Quote into a simple array containing payment lines.

Notes:

* Completed implementation of Abstract conversion logic.
* Completed implementation of Quote conversion logic.
* Fixed a bunch of notices in several files (missing type hints, global function calls without import statements etc.).
* Patched unit tests in accordance with the changes applied.

+++ Model/Api/Payment/Converter/Item/ItemInterface.php
@@ -56,9 +56,9 @@ interface ItemInterface
     public function getUnitAmountWithoutVat(): float;
 
     /**
-     * @return float
+     * @return int
      */
-    public function getVatPct(): float;
+    public function getVatPct(): int;
 
     /**
      * @return string
